fix: ApplyScreen 로직 useHooks로 분리
