+++ package.json
@@ -11,6 +11,7 @@
     "clean-ios": "rm -rf ~/Library/Developer/Xcode/DerivedData && cd ios && rm -rf Podfile.lock && pod update && pod install"
   },
   "dependencies": {
+    "@react-native-community/netinfo": "^11.2.1",
     "@react-native-firebase/app": "^18.8.0",
     "@react-native-firebase/auth": "^18.8.0",
     "@react-navigation/native": "^6.1.9",
