+++ src/screens/ApplyScreen/ApplyScreen.tsx
@@ -1,99 +1,22 @@
-import React, {useEffect, useState} from 'react';
+import React from 'react';
 import {SafeAreaProvider, SafeAreaView} from 'react-native-safe-area-context';
-import {Alert, FlatList, Platform, StyleSheet} from 'react-native';
+import {FlatList, Platform, StyleSheet} from 'react-native';
 import {KeyboardAvoidingView} from 'native-base';
 import Header from '@/components/ApplyScreen/Header';
 import Footer from '@/components/ApplyScreen/Footer';
 import TodoEmpty from '@/components/Empty/TodoEmpty';
 import TodoCard from '@/components/Card/TodoCard';
-import Realm from 'realm';
-import dayjs from 'dayjs';
-import {TodoTestSchema} from '@/realm/TodoTestSchema';
-
-type TodoListType = {
-  id: string;
-  title: string;
-  isChecked: false;
-  create_at: string;
-};
+import useApplyScreenHooks from '@/useApplyScreenHooks';
 
 const ApplyScreen = () => {
-  const [todoList, setTodoList] = useState<TodoListType[]>([]);
-  const [inputTodo, setInputTodo] = useState('');
-
-  // *legacy
-  // const [selectedTodo, setSelectedTodo] = useState<string[]>([]);
-  // const onPressCheck = (idx: string) => {
-  //   setSelectedTodo(prev => {
-  //     const result = prev.includes(idx)
-  //       ? prev.filter(i => i !== idx)
-  //       : [idx, ...prev];
-  //     return result;
-  //   });
-  // };
-
-  const onSubmit = () => {
-    if (inputTodo.length > 0) {
-      setInputTodo('');
-      // realm todo create
-      createDB(inputTodo);
-    } else {
-      Alert.alert('알림', '할일을 적어주세요!', [
-        {text: '확인', onPress: () => {}},
-      ]);
-    }
-  };
-
-  // ==================
-  // ===[Realm Test]===
-  // ==================
-  // * relam config
-  // The initial schemaVersion is 0.
-  const realm = new Realm({
-    schema: [TodoTestSchema],
-  });
-  const readTodo = () => {
-    const result = realm.objects('todotest').toJSON() as TodoListType[];
-    setTodoList(result);
-    // console.log(result);
-  };
-  const createDB = (title: string) => {
-    const id = new Realm.BSON.UUID();
-    realm.write(() => {
-      realm.create('todotest', {
-        id: id.toString(),
-        title,
-        isChecked: false,
-        create_at: dayjs().toString(),
-      });
-    });
-    readTodo();
-  };
-  const deleteTodoList = (id: string) => {
-    const deleteDB = realm.objects('todotest').filtered(`id = '${id}'`)[0];
-    realm.write(() => {
-      realm.delete(deleteDB);
-    });
-
-    // 삭제 후 데이터 갱신
-    readTodo();
-  };
-  const updateTodoList = (id: string) => {
-    const targetData = realm.objects('todotest').filtered(`id = '${id}'`)[0];
-    // console.log('[targetData] : ', targetData);
-
-    realm.write(() => {
-      targetData.isChecked = !targetData.isChecked;
-    });
-
-    // 수정 후 데이터 갱신
-    readTodo();
-  };
-
-  useEffect(() => {
-    readTodo();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  const {
+    inputTodo,
+    setInputTodo,
+    todoList,
+    deleteTodoList,
+    onSubmit,
+    updateTodoList,
+  } = useApplyScreenHooks();
 
   // Issue : TextInput이 하단배치시에 헤더고정이 안되어 native 설정 변경
   // AndroidManifest.xml에서
